fix: variables not updated on var manager
