+++ apps/web/src/components/templates/VariableManager.cy.tsx
@@ -3,12 +3,15 @@ import { useFormContext } from 'react-hook-form';
 import { TestWrapper } from '../../testing';
 import { VariableManager } from './VariableManager';
 import { TemplateFormProvider } from './TemplateFormProvider';
+import { useVariablesManager } from '../../hooks/use-variables-manager';
 
 it('should show available variables - string', function () {
+  const variablesArray = useVariablesManager(0, ['content']);
+
   mount(
     <TestWrapper>
       <TemplateFormProvider>
-        <VariableManager index={0} contents={['content']} />
+        <VariableManager index={0} variablesArray={variablesArray} />
         <FormTester content={'Hello, {{ name }}'} />
       </TemplateFormProvider>
     </TestWrapper>
@@ -20,10 +23,12 @@ it('should show available variables - string', function () {
 });
 
 it('should show available variables - array', function () {
+  const variablesArray = useVariablesManager(0, ['content']);
+
   mount(
     <TestWrapper>
       <TemplateFormProvider>
-        <VariableManager index={0} contents={['content']} />
+        <VariableManager index={0} variablesArray={variablesArray} />
         <FormTester content={'Hello, {{#each name}} {{/each}}'} />
       </TemplateFormProvider>
     </TestWrapper>
@@ -35,10 +40,12 @@ it('should show available variables - array', function () {
 });
 
 it('should show available variables including nested - array', function () {
+  const variablesArray = useVariablesManager(0, ['content']);
+
   mount(
     <TestWrapper>
       <TemplateFormProvider>
-        <VariableManager index={0} contents={['content']} />
+        <VariableManager index={0} variablesArray={variablesArray} />
         <FormTester content={'Hello, {{#each name}} {{nested_variable}} {{/each}}'} />
       </TemplateFormProvider>
     </TestWrapper>
@@ -52,10 +59,12 @@ it('should show available variables including nested - array', function () {
 });
 
 it('should show available variables - boolean', function () {
+  const variablesArray = useVariablesManager(0, ['content']);
+
   mount(
     <TestWrapper>
       <TemplateFormProvider>
-        <VariableManager index={0} contents={['content']} />
+        <VariableManager index={0} variablesArray={variablesArray} />
         <FormTester content={'Hello, {{#if name}} {{/if}}'} />
       </TemplateFormProvider>
     </TestWrapper>
@@ -67,10 +76,12 @@ it('should show available variables - boolean', function () {
 });
 
 it('should show available variables including nested - boolean', function () {
+  const variablesArray = useVariablesManager(0, ['content']);
+
   mount(
     <TestWrapper>
       <TemplateFormProvider>
-        <VariableManager index={0} contents={['content']} />
+        <VariableManager index={0} variablesArray={variablesArray} />
         <FormTester content={'Hello, {{#if name}} {{nested_variable}} {{/if}}'} />
       </TemplateFormProvider>
     </TestWrapper>
@@ -84,10 +95,12 @@ it('should show available variables including nested - boolean', function () {
 });
 
 it('should show available variables including deeply nested', function () {
+  const variablesArray = useVariablesManager(0, ['content']);
+
   mount(
     <TestWrapper>
       <TemplateFormProvider>
-        <VariableManager index={0} contents={['content']} />
+        <VariableManager index={0} variablesArray={variablesArray} />
         <FormTester
           content={
             'Hello, {{#if name}} {{nested_variable}} {{#each nested_name}} {{deeply_nested_variable}} {{/each}} {{/if}}'
@@ -109,10 +122,12 @@ it('should show available variables including deeply nested', function () {
 });
 
 it('should show reserved variables', function () {
+  const variablesArray = useVariablesManager(0, ['content']);
+
   mount(
     <TestWrapper>
       <TemplateFormProvider>
-        <VariableManager index={0} contents={['content']} />
+        <VariableManager index={0} variablesArray={variablesArray} />
         <FormTester content={'Hello, {{ subscriber }}'} />
       </TemplateFormProvider>
     </TestWrapper>
