+++ apps/web/src/components/templates/email-editor/EmailMessagesCards.tsx
@@ -12,6 +12,7 @@ import { WebIcon } from '../../../pages/templates/editor/PreviewSegment/WebIcon'
 import { useHotkeys } from '@mantine/hooks';
 import { VariableManager } from '../VariableManager';
 import { VariablesManagement } from './variables-management/VariablesManagement';
+import { useVariablesManager } from '../../../hooks/use-variables-manager';
 
 export enum ViewEnum {
   EDIT = 'Edit',
@@ -25,6 +26,7 @@ export function EmailMessagesCards({ index, isIntegrationActive }: { index: numb
   const [preview, setPreview] = useState<'mobile' | 'web'>('web');
   const theme = useMantineTheme();
   const [modalOpen, setModalOpen] = useState(false);
+  const variablesArray = useVariablesManager(index, ['content', 'htmlContent', 'subject']);
 
   useHotkeys([
     [
@@ -71,6 +73,7 @@ export function EmailMessagesCards({ index, isIntegrationActive }: { index: numb
           <Grid.Col p={0} span={6}>
             <When truthy={view === ViewEnum.PREVIEW}>
               <SegmentedControl
+                data-test-id="preview-mode-switch"
                 styles={{
                   root: {
                     background: 'transparent',
@@ -165,9 +168,10 @@ export function EmailMessagesCards({ index, isIntegrationActive }: { index: numb
         centered
         overflow="inside"
       >
-        <VariableManager hideLabel={true} index={index} contents={['content', 'htmlContent', 'subject']} />
+        <VariableManager hideLabel={true} index={index} variablesArray={variablesArray} />
         <Group position="right">
           <Button
+            data-test-id="close-var-manager-modal"
             mt={30}
             onClick={() => {
               setModalOpen(false);
